requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
youtube-transcript-api>=0.6.2
httpx==0.25.2
python-multipart==0.0.6
python-dotenv==1.0.0